GitHub Deployments: Fix target dir prefix on input
