+++ client/my-sites/github-deployments/components/github-connection-form/index.tsx
@@ -4,8 +4,6 @@ import { useI18n } from '@wordpress/react-i18n';
 import { ChangeEvent, useMemo, useRef, useState } from 'react';
 import FormFieldset from 'calypso/components/forms/form-fieldset';
 import FormSelect from 'calypso/components/forms/form-select';
-import FormSettingExplanation from 'calypso/components/forms/form-setting-explanation';
-import FormTextInput from 'calypso/components/forms/form-text-input';
 import { GitHubInstallationData } from 'calypso/my-sites/github-deployments/use-github-installations-query';
 import { useGithubRepositoryBranchesQuery } from 'calypso/my-sites/github-deployments/use-github-repository-branches-query';
 import { useGithubRepositoryChecksQuery } from 'calypso/my-sites/github-deployments/use-github-repository-checks-query';
@@ -13,6 +11,7 @@ import { GitHubRepositoryData } from '../../use-github-repositories-query';
 import { AutomatedDeploymentsToggle } from '../automated-deployments-toggle';
 import { DeploymentStyle } from '../deployment-style';
 import { useCheckWorkflowQuery } from '../deployment-style/use-check-workflow-query';
+import { TargetDirInput } from '../target-dir-input';
 
 import './style.scss';
 
@@ -167,22 +166,7 @@ export const GitHubConnectionForm = ( {
 						{ isFetchingBranches && <Spinner /> }
 					</div>
 				</FormFieldset>
-				<FormFieldset>
-					<FormLabel htmlFor="target">{ __( 'Destination directory' ) }</FormLabel>
-					<FormTextInput
-						id="target"
-						value={ destPath }
-						onChange={ ( event: ChangeEvent< HTMLInputElement > ) => {
-							let targetDir = event.currentTarget.value.trim();
-							targetDir = targetDir.startsWith( '/' ) ? targetDir : `/${ targetDir }`;
-
-							setDestPath( targetDir );
-						} }
-					/>
-					<FormSettingExplanation>
-						{ __( 'This path is relative to the server root' ) }
-					</FormSettingExplanation>
-				</FormFieldset>
+				<TargetDirInput onChange={ setDestPath } value={ destPath } />
 				<AutomatedDeploymentsToggle
 					onChange={ setIsAutoDeploy }
 					value={ isAutoDeploy }
